Architecture improved. Moved cmd to the root.
The full pipeline is set up in pusher/pusher, and the command has been
moved to the root. In this way we reflect the plan to make this produce
a single binary `pusher`
