+++ fileinfo/fileinfo.go
@@ -8,7 +8,7 @@ import (
 
 // A LocalDataFile holds all the information we require about a file.
 type LocalDataFile struct {
-	FullRelativeName string
+	AbsoluteFileName  string
 	Info             os.FileInfo
 	// Cache the size so sums of filesize are always using the same
 	// integer value, in case the file grows in size between receipt and
