+++ pusher.go
@@ -1,35 +1,43 @@
-package pusher
+package main
 
 import (
+	"flag"
+	"log"
 	"time"
+
 	"github.com/m-lab/pusher/bytecount"
-	"github.com/m-lab/pusher/fileinfo"
-	"github.com/m-lab/pusher/tarcache"
-	"github.com/m-lab/pusher/uploader"
+	"github.com/m-lab/pusher/finder"
+	"github.com/m-lab/pusher/pusher"
 )
 
+var (
+	directory     = flag.String("directory", "/var/spool/test", "The directory to watch for files.")
+	project       = flag.String("project", "mlab-sandbox", "The GCP project for uploading")
+	bucket        = flag.String("bucket", "dropbox-mlab-sandbox", "The GCP bucket for uploading")
+	ageThreshold    = flag.Duration("age_threshold", time.Duration(2) * time.Hour, "The maximum amount of time we should hold onto a piece of data before uploading it.")
+	sizeThreshold = bytecount.ByteCount(20 * bytecount.Megabyte)
+)
 
-// Config holds all config information needed for a new pusher channel.
-type Config struct {
-	Project, Bucket, Creds string
-	TarfileSizeThreshold bytecount.ByteCount
-	FileAgeThreshold time.Duration
+func init() {
+	flag.Var(&sizeThreshold, "size_threshold", "The minimum tarfile size we require to commence upload (1KB, 200MB, etc). Default is 20MB")
 }
 
-// New creates a new channel to which files for upload should be sent, and also
-// sets up the goroutine which does the tar and push operations.
-func New(config Config) chan *fileinfo.LocalDataFile {
-	// Set up the processing chain.
-	uploader := uploader.New(config.Project, config.Bucket, config.Creds)
-	tarCache := tarcache.New(config.TarfileSizeThreshold, config.FileAgeThreshold, uploader)
-
-	// By giving this channel a large buffer, we attempt to decouple file
-	// discovery event response times from any file processing times.
-	c := make(chan *fileinfo.LocalDataFile, 1000000)
-	go func() {
-		for file := range c {
-			tarCache.Add(file)
+func main() {
+	flag.Parse()
+	config := pusher.Config{
+		Project: *project,
+		Bucket: *bucket,
+		TarfileSizeThreshold: sizeThreshold,
+		FileAgeThreshold: *ageThreshold,
+	}
+	pusherChannel := pusher.New(config)
+	interval := time.Duration(10) * time.Minute
+	for {
+		files, err := finder.FindFiles(*directory, interval)
+		log.Println(files, err)
+		for _, file := range files {
+			pusherChannel <- file
 		}
-	}()
-	return c
+		time.Sleep(interval)
+	}
 }
